+++ src/constants/sceneAssetList.js
@@ -1,0 +1,63 @@
+export const SCENE_ASSET_LIST = [
+  {
+    type: 'coster',
+    name: 'coster',
+    src: '/images/coster.png',
+  },
+  {
+    type: 'track',
+    name: 'steel',
+    src: '/images/steel.png',
+  },
+  {
+    type: 'track',
+    name: 'stone',
+    src: '/images/stone.png',
+  },
+  {
+    type: 'track',
+    name: 'vine',
+    src: '/images/vine.png',
+  },
+  {
+    type: 'track',
+    name: 'slime',
+    src: '/images/slime.png',
+  },
+  {
+    type: 'item',
+    name: 'flying-swings',
+    src: '/images/flying-swings.png',
+    model: '/objects/flying-swings.glb',
+  },
+  {
+    type: 'item',
+    name: 'bench',
+    src: '/images/bench.png',
+    model: '/objects/bench.glb',
+  },
+  {
+    type: 'item',
+    name: 'iceCreamCar',
+    src: '/images/ice-cream-car.png',
+    model: '/objects/ice-cream-car.glb',
+  },
+  {
+    type: 'item',
+    name: 'bumperCar',
+    src: '/images/bumper-car.png',
+    model: '/objects/bumper-car.glb',
+  },
+  {
+    type: 'item',
+    name: 'ferrisWheel',
+    src: '/images/ferris-wheel.png',
+    model: '/objects/ferris-wheel.glb',
+  },
+  {
+    type: 'item',
+    name: 'coffeeCart',
+    src: '/images/coffee-cart.png',
+    model: '/objects/coffee-cart.glb',
+  },
+];
